CV Control - Face targeting
Since facial recognition is currently more simple to do for creating a first test, this program now targets the center of a face and points a vector from the center of the face to the center of the screen. This will later be used to vector the turret into firing position.